repair bug fixing front end

+++ server/api/models/stockAddingLogs.mode.js
@@ -1,5 +1,5 @@
 const createInventoryLog = async (db) => {
-    await db.query(`
+  await db.query(`
     CREATE TABLE IF NOT EXISTS stock_order_log (
       id INT AUTO_INCREMENT PRIMARY KEY,
       invoiceid VARCHAR(50) UNIQUE,
@@ -12,25 +12,29 @@ const createInventoryLog = async (db) => {
       FOREIGN KEY (product_id) REFERENCES products(id)
       )
     `);
-  };
+};
 
-  const alterInventoryLog = async (db) => {
-    try {
-      
+const alterInventoryLog = async (db) => {
+  try {
+    const [columns] = await db.query(`
+    SHOW COLUMNS FROM stock_order_log LIKE 'invoiceid';
+  `);
+
+    if (columns.length > 0) {
       await db.query(`
         ALTER TABLE stock_order_log
         DROP INDEX invoiceid;
       `);
-  
-      await db.query(`
-        ALTER TABLE stock_order_log
-        DROP COLUMN invoiceid;
-      `);
-  
-      console.log("Invoice ID column and its unique constraint dropped successfully.");
-    } catch (error) {
-      console.error("Error altering the table:", error);
     }
-  };
+
   
-  module.exports = { createInventoryLog, alterInventoryLog };
+
+    console.log(
+      "Invoice ID column and its unique constraint dropped successfully."
+    );
+  } catch (error) {
+    console.error("Error altering the table:", error);
+  }
+};
+
+module.exports = { createInventoryLog, alterInventoryLog };
